libosdp: Make osdp_cmd_mfg and osdp_mfgrep identical
As pointed out by adakus@ in zephyr review, we may want to allow PDs to
send unsolicited MFGREP responses in future. To allow for that without
having to deal with it as a bug, let us fix this one proactively.

This commit also fixes a typo and bumps utils submodule to latest
version.

Link: https://github.com/zephyrproject-rtos/zephyr/pull/46449

+++ include/osdp.h
@@ -622,8 +622,8 @@ struct osdp_event_keypress {
  */
 struct osdp_event_mfgrep {
 	uint32_t vendor_code;
-	int command;
-	int length;
+	uint8_t command;
+	uint8_t length;
 	uint8_t data[OSDP_EVENT_MAX_DATALEN];
 };
